+++ src/main/scala/edu/ie3/simona/config/SimonaConfig.scala
@@ -108,6 +108,8 @@ object SimonaConfig {
   ) extends GridConfigParams
 
   final case class Simona(
+      congestionManagement: Simona.CongestionManagement =
+        Simona.CongestionManagement(),
       control: Option[Simona.Control] = None,
       gridConfig: Simona.GridConfig = Simona.GridConfig(),
       input: InputConfig,
@@ -118,6 +120,11 @@ object SimonaConfig {
       time: Simona.Time = Simona.Time(),
   )
   object Simona {
+    final case class CongestionManagement(
+        enableDetection: Boolean = false,
+        timeout: FiniteDuration = 30.seconds,
+    )
+
     final case class Control(
         transformer: List[TransformerControlGroup] = List.empty
     )