+++ src/test/scala/edu/ie3/simona/model/grid/Transformer3wModelSpec.scala
@@ -16,6 +16,11 @@ import edu.ie3.simona.model.grid.Transformer3wPowerFlowCase.{
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.simona.test.common.input.Transformer3wTestData
 import edu.ie3.util.quantities.PowerSystemUnits._
+import edu.ie3.util.scala.quantities.{
+  ApparentPower,
+  Megavoltamperes,
+  Voltamperes,
+}
 import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor4}
 import squants.Each
 import tech.units.indriya.quantity.Quantities
@@ -28,6 +33,7 @@ class Transformer3wModelSpec
     with Transformer3wTestData {
   val testingTolerance = 1e-5
   implicit val dimensionlessTolerance: squants.Dimensionless = Each(1e-8)
+  implicit val powerTolerance: ApparentPower = Voltamperes(1e-3)
 
   "A three winding transformer input model" should {
     "be validated without an exception from a valid input model" in {
@@ -67,6 +73,7 @@ class Transformer3wModelSpec
               transformerTappingModel,
               amount,
               powerFlowCase,
+              sRated,
               r,
               x,
               g,
@@ -85,6 +92,7 @@ class Transformer3wModelSpec
           transformerTappingModel shouldBe expectedTappingModel
           amount shouldBe transformer3wInput.getParallelDevices
           powerFlowCase shouldBe PowerFlowCaseA
+          sRated should approximate(Megavoltamperes(120))
           r should approximate(Each(1.03878e-3))
           x should approximate(Each(166.34349e-3))
           g should approximate(Each(1.874312e-6))
@@ -142,6 +150,7 @@ class Transformer3wModelSpec
               transformerTappingModel,
               amount,
               powerFlowCase,
+              sRated,
               r,
               x,
               g,
@@ -160,6 +169,7 @@ class Transformer3wModelSpec
           transformerTappingModel shouldBe expectedTappingModel
           amount shouldBe transformer3wInput.getParallelDevices
           powerFlowCase shouldBe PowerFlowCaseB
+          sRated should approximate(Megavoltamperes(60))
           r should approximate(Each(240.9972299e-6))
           x should approximate(Each(24.99307479224e-3))
           g should approximate(Each(0d))
@@ -217,6 +227,7 @@ class Transformer3wModelSpec
               transformerTappingModel,
               amount,
               powerFlowCase,
+              sRated,
               r,
               x,
               g,
@@ -235,6 +246,7 @@ class Transformer3wModelSpec
           transformerTappingModel shouldBe expectedTappingModel
           amount shouldBe transformer3wInput.getParallelDevices
           powerFlowCase shouldBe PowerFlowCaseC
+          sRated should approximate(Megavoltamperes(40))
           r should approximate(Each(3.185595567e-6))
           x should approximate(Each(556.0941828e-6))
           g should approximate(Each(0d))